Move external formats to formats app
See #1888
