+++ weblate/formats/external.py
